common/rc: Enable _format_swapfile to return the swap size
Once the kernel is free to not map the full swap file during a swapon call,
it can be useful to know the exact size of the swap area created during
_format_swapfile().

To achieve this, it is required to change _require_scratch_swapfile(), to drop
the _format_swapfile() return value, otherwise, it will also have a return value
that will end up in tests outputs causing tests to fail.

Test currently using _format_swapfile() do not need to be updated, the
function's returned value will simply be ignored by default.

V2:	- return swap size in bytes
	- add a function description to _format_swapfile
	- rephrase patch's description